// web/src/gen/model/nodePaletteItemDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { IReadOnlyListOfPinDefinition } from "./iReadOnlyListOfPinDefinition";

export interface NodePaletteItemDto {
  key: string;
  label: string;
  category: string;
  source: string;
  executor: string;
  inputs: IReadOnlyListOfPinDefinition;
  outputs: IReadOnlyListOfPinDefinition;
  /** @nullable */
  id?: string | null;
}